requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
flask
langchain 
PyPDF2
pypdf
python-dotenv
langchain-community
langchain-openai
langchain-pinecone
flask-session
-e .